# openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Точка опоры — API мини-приложения и каталога помощи",
    "version": "submission-2026-09-28",
    "description": "Публичный каталог помощи и отдельные сеансы MAX. Редактор не входит в публичный API. Тексты и доступность услуг требуют уточнения у организации. Источники вакансий и жилья могут быть недоступны."
  },
  "paths": {
...
    "/api/session": {
      "post": {
        "summary": "Launch",
        "operationId": "launch_api_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Launch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "session"
                  ],
                  "properties": {
                    "session": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Неверные или просроченные init_data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Повторный запуск раньше 5 секунд",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Лимит 200 сеансов",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Чужой Origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Тело больше 24000 байт",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/favorites": {
      "get": {
        "summary": "Favorites",
        "operationId": "favorites_api_favorites_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "cards"
                  ],
                  "properties": {
                    "cards": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Card"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Сеанс завершён",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "MaxSession": []
          }
        ]
      },
      "post": {
        "summary": "Favorite",
        "operationId": "favorite_api_favorites_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FavoriteAction"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "cards"
                  ],
                  "properties": {
                    "cards": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Card"
                      }
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Сеанс завершён",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Материал снят",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Карточка не открывалась в этом сеансе",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Чужой Origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Тело больше 24000 байт",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "MaxSession": []
          }
        ]
      }
    },
    "/api/material": {
      "post": {
        "summary": "Open Material",
        "operationId": "open_material_api_material_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaterialAction"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "card",
                    "saved"
                  ],
                  "properties": {
                    "card": {
                      "$ref": "#/components/schemas/Card"
                    },
                    "saved": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Сеанс завершён",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Материал недоступен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Чужой Origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Тело больше 24000 байт",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "MaxSession": []
          }
        ],
        "description": "save=false только открывает ответ. save=true сохраняет в Моё. Разрешены только материалы явного реестра saved_answers.json."
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FavoriteAction": {
        "properties": {
          "card_id": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$",
            "title": "Card Id"
          },
          "saved": {
            "type": "boolean",
            "title": "Saved"
          }
        },
        "type": "object",
        "required": [
          "card_id",
          "saved"
        ],
        "title": "FavoriteAction"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Launch": {
        "properties": {
          "init_data": {
            "type": "string",
            "maxLength": 16000,
            "title": "Init Data",
            "default": ""
          }
        },
        "type": "object",
        "title": "Launch"
      },
      "MaterialAction": {
        "properties": {
          "node_id": {
            "type": "string",
            "pattern": "^[A-Za-z][A-Za-z0-9_-]{0,79}$",
            "title": "Node Id"
          },
          "save": {
            "type": "boolean",
            "title": "Save",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "node_id"
        ],
        "title": "MaterialAction"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "Error": {
        "type": "object",
        "required": [
          "detail"
        ],
        "properties": {
          "detail": {}
        }
      },
...
      "Card": {
        "type": "object",
        "required": [
          "text"
        ],
        "properties": {
          "text": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "section": {
            "type": "string"
          },
          "favorite_id": {
            "type": "string"
          },
          "node_id": {
            "type": "string"
          },
          "actions": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "saved_at": {
            "type": "string"
          },
          "unavailable": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "MaxSession": {
        "type": "http",
        "scheme": "bearer",
        "description": "Токен из POST /api/session с подписанным MAX init_data. До 1 часа бездействия; исчезает при перезапуске. Пароль MAX и токен бота не передаются."
      }
    }
  },
  "servers": [
    {
      "url": "https://135-106-229-210.sslip.io"
    }
  ]
}